% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_aes.R
\name{calc_ContrastColor}
\alias{calc_ContrastColor}
\title{Calculate the contrast color for a given color.}
\usage{
calc_ContrastColor(input_color)
}
\arguments{
\item{input_color}{ESPNCricinfo Player ID.}
}
\value{
"black" or "white" depending on contrast of \code{input_color}
}
\description{
This function takes in a color (\code{input_color}) returns a color in contrast to the input color.
}
\examples{
calc_ContrastColor("white")
calc_ContrastColor("red")
calc_ContrastColor("blue")
calc_ContrastColor("black")
}
\keyword{Misc}
