% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismissal.r
\name{dispBatsmanDismissalsByBowlerType}
\alias{dispBatsmanDismissalsByBowlerType}
\title{Plots a pie chart of Dismissal Summary vs type and style of Bowlers}
\usage{
dispBatsmanDismissalsByBowlerType(data)
}
\arguments{
\item{data}{Output of splitDismissalSummary().}
}
\value{
Plots a pie chart of Dismissal Summary vs type and style of Bowlers
}
\description{
This function takes in output of splitDismissalSummary  and Plots a pie chart of Dismissal Summary vs type and style of Bowlers.
}
\examples{
sachin = getDismissalSummary(35320,11)
sachin1 = splitDismissalSummary(sachin)
dispBatsmanDismissalsByBowlerType(sachin1)
}
