% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{dispBattingAveByPosPlayed}
\alias{dispBattingAveByPosPlayed}
\title{Displays Barplot of Players's Batting according to the Player's Batting Position}
\usage{
dispBattingAveByPosPlayed(
  data,
  plot_title = "Batting Average by Position Played"
)
}
\arguments{
\item{data}{Output of splitBattingSummary.}
}
\value{
Plots a player's batting position-wise batting average.
}
\description{
This function takes in "Batting Summary" dataframe modified after splitBattingSummary and plots a player's batting position-wise batting average.
}
\examples{
sachin_BattingSummary = statsguRu::get_Summary_Batting(PlayerID = 35320, MatchType = 2)
sachin_BattingSummary = statsguRu::split_Summary_Batting(sachin_BattingSummary, MatchType = 2)
statsguRu::dispBattingAveByPosPlayed(data = sachin_BattingSummary)
}
\keyword{Batting}
