% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{dispBattingAveByYears}
\alias{dispBattingAveByYears}
\title{Displays Barplot of Players's Batting Average Year-wise}
\usage{
dispBattingAveByYears(data, plot_title = "Year Wise Batting Average")
}
\arguments{
\item{data}{Output of splitBattingSummary.}
}
\value{
Plots a player's year-wise batting average.
}
\description{
This function takes in "Batting Summary" dataframe modified after splitBattingSummary and plots a player's year-wise batting average.
}
\examples{
sachin_BattingSummary = statsguRu::get_Summary_Batting(PlayerID = 35320, MatchType = 2)
sachin_BattingSummary = statsguRu::split_Summary_Batting(sachin_BattingSummary, MatchType = 2)
statsguRu::dispBattingAveByYears(data = sachin_BattingSummary)
}
\keyword{Batting}
