% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting_dismissals.r
\name{dispBattingDismissalsByBowlerType}
\alias{dispBattingDismissalsByBowlerType}
\title{Plots a pie chart of Dismissal Summary vs type and style of Bowlers}
\usage{
dispBattingDismissalsByBowlerType(
  data,
  plot_title = "Distribution of Dismissals vs Bowlers"
)
}
\arguments{
\item{data}{Output of statsguRu::split_DismissalSummary_Batting().}
}
\value{
Plots a pie chart of Dismissal Summary vs type and style of Bowlers
}
\description{
This function takes in output of splitDismissalSummary  and Plots a pie chart of Dismissal Summary vs type and style of Bowlers.
}
\examples{
sachin_DismissalSummary_Batting = statsguRu::get_DismissalSummary_Batting(PlayerID = 35320, MatchType = 2)
sachin_DismissalSummary_Batting = statsguRu::split_DismissalSummary_Batting(sachin_DismissalSummary_Batting)
statsguRu::dispBattingDismissalsByBowlerType(data = sachin_DismissalSummary_Batting)
}
\keyword{Batting}
