% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowling.R
\name{dispBowlingERByYears}
\alias{dispBowlingERByYears}
\title{Displays Barplot of Players's Year-wise Bowling Economy Rate}
\usage{
dispBowlingERByYears(data, plot_title = "Year Wise Bowling Economy Rate")
}
\arguments{
\item{data}{Output of statsguRu::split_Summary_Bowling()}
}
\value{
Plots a player's year-wise Bowling Economy Rate
}
\description{
This function takes in "Batting Summary" dataframe modified after splitBowlingSummary and plots a player's Year-wise Bowling Economy Rate.
}
\examples{
ashwin_BowlingSummary = statsguRu::get_Summary_Bowling(PlayerID = 26421, MatchType = 1)
ashwin_BowlingSummary = statsguRu::split_Summary_Bowling(data = ashwin_BowlingSummary, MatchType = 1)
statsguRu::dispBowlingERByYears(data = ashwin_BowlingSummary)
}
\keyword{Bowling}
