% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting_dismissals.r
\name{dispDismissalsBatting}
\alias{dispDismissalsBatting}
\title{Plots a pie chart of Dismissal Summary of a Player}
\usage{
dispDismissalsBatting(data, plot_title = "Distribution of Dismissals")
}
\arguments{
\item{data}{Output of splitDismissalSummary().}
}
\value{
Plots a pie chart of Dismissal Summary of a player.
}
\description{
This function takes in output of splitDismissalSummary and plots the Dismissal Summary of a Player as a pie chart.
}
\examples{
sachin_DismissalSummary_Batting = statsguRu::get_DismissalSummary_Batting(PlayerID = 35320, MatchType = 2)
sachin_DismissalSummary_Batting = statsguRu::split_DismissalSummary_Batting(sachin_DismissalSummary_Batting)
statsguRu::dispDismissalsBatting(data = sachin_DismissalSummary_Batting)
}
\keyword{Batting}
