% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting_dismissals.r
\name{get_DismissalSummary_Batting}
\alias{get_DismissalSummary_Batting}
\title{Gets a Players's Dismissal Summary (whilst Batting)}
\usage{
get_DismissalSummary_Batting(PlayerID, MatchType)
}
\arguments{
\item{PlayerID}{ESPNCricinfo Player ID.}

\item{MatchType}{Type of Match Played (1 for Test; 2 for ODI; 3 for T20I ; 11 for All)}
}
\value{
Returns DataFrame containing complete Batting Dismissals Summary of a Player
}
\description{
This function scraps exclusive dismissal summary (as a batsman) info from Statsguru and returns a dataframe
}
\examples{
sachin_DismissalSummary_Batting = statsguRu::get_DismissalSummary_Batting(PlayerID = 35320, MatchType = 2)
print(sachin_DismissalSummary_Batting)
}
\keyword{Batting}
