% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowling_Dismissals.R
\name{get_DismissalSummary_Bowling}
\alias{get_DismissalSummary_Bowling}
\title{Gets a Players's Wickets Summary (whilst Bowling)}
\usage{
get_DismissalSummary_Bowling(PlayerID, MatchType)
}
\arguments{
\item{PlayerID}{ESPNCricinfo Player ID.}

\item{MatchType}{Type of Match Played (1 for Test; 2 for ODI; 3 for T20I ; 11 for All)}
}
\value{
Returns DataFrame containing complete Wickets Summary of a Bowler
}
\description{
This function scraps exclusive dismissal summary (as a bowler) info from Statsguru and returns a dataframe
}
\examples{
ashwin_DismissalSummary_Bowling = statsguRu::get_DismissalSummary_Bowling(PlayerID = 26421, MatchType = 1)
print(ashwin_DismissalSummary_Bowling)
}
\keyword{Bowling}
