% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{get_Summary_Batting}
\alias{get_Summary_Batting}
\title{Gets a Players's Batting Summary}
\usage{
get_Summary_Batting(PlayerID, MatchType)
}
\arguments{
\item{PlayerID}{ESPNCricinfo Player ID.}

\item{MatchType}{Type of Match Played (1 for Test; 2 for ODI; 3 for T20I ; 11 for All)}
}
\value{
Returns dataframe containing complete Batting Summary of a Player
}
\description{
This function scraps batting summary info from Statsguru and returns a dataframe
}
\examples{
sachin_BattingSummary = statsguRu::get_Summary_Batting(PlayerID = 35320, MatchType = 11)
print(sachin_BattingSummary)
}
\keyword{Batting}
