% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{get_Summary_Batting_Overall}
\alias{get_Summary_Batting_Overall}
\title{Gets a Players's Overall Batting Summary}
\usage{
get_Summary_Batting_Overall(PlayerID, MatchType)
}
\arguments{
\item{PlayerID}{ESPNCricinfo Player ID.}

\item{MatchType}{Type of Match Played (1 for Test; 2 for ODI; 3 for T20I ; 11 for All)}
}
\value{
Returns dataframe containing overall Batting Summary of a Player
}
\description{
This function scraps overall batting summary info from Statsguru and returns the data as a dataframe.
}
\examples{
sachin_BattingSummary_Overall = statsguRu::get_Summary_Batting_Overall(PlayerID = 35320, MatchType = 11)
print(sachin_BattingSummary_Overall)
}
\keyword{Batting}
