% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowling.R
\name{getBowlingSummary}
\alias{getBowlingSummary}
\title{Gets a Players's Bowling Summary}
\usage{
getBowlingSummary(PlayerID, MatchType)
}
\arguments{
\item{PlayerID}{ESPNCricinfo Player ID.}

\item{MatchType}{Type of Match Played (1 for Test; 2 for ODI; 3 for T20I ; 11 for All)}
}
\value{
Returns dataframe containing complete Batting Summary of a Player
}
\description{
This function scraps bowling summary info from Statsguru and returns a dataframe
}
\examples{
sachin = getBowlingSummary(35320,11)
}
