% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismissal.r
\name{getDismissalSummary}
\alias{getDismissalSummary}
\title{Gets a Players's Dismissal Summary (whilst Batting)}
\usage{
getDismissalSummary(PlayerID, MatchType)
}
\arguments{
\item{PlayerID}{ESPNCricinfo Player ID.}

\item{MatchType}{Type of Match Played (1 for Test; 2 for ODI; 3 for T20I ; 11 for All)}
}
\value{
Returns DataFrame containing complete Batting Dismissals Summary of a Player
}
\description{
This function scraps exclusive dismissal summary (as a batsman) info from Statsguru and returns a dataframe
}
\examples{
sachin = getDismissalSummary(35320,11)
}
