% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{getOverallBattingSummary}
\alias{getOverallBattingSummary}
\title{Gets a Players's Overall Batting Summary}
\usage{
getOverallBattingSummary(PlayerID, MatchType)
}
\arguments{
\item{PlayerID}{ESPNCricinfo Player ID.}

\item{MatchType}{Type of Match Played (1 for Test; 2 for ODI; 3 for T20I ; 11 for All)}
}
\value{
Returns dataframe containing overall Batting Summary of a Player
}
\description{
This function scraps overall batting summary info from Statsguru and returns a dataframe
}
\examples{
sachin = getBattingSummary(35320,11)
}
