% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_aes.R
\name{getTeamColors}
\alias{getTeamColors}
\title{Get Team Jersey Colors}
\usage{
getTeamColors()
}
\value{
A named vector containing hex code of jersey colors.
}
\description{
This function returns a named vector containing hex code of all national team's jersey.
}
\examples{
statsguRu::getTeamColors()
}
\keyword{Misc}
