% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlDismissal.R
\name{getWicketsSummary}
\alias{getWicketsSummary}
\title{Gets a Players's Wickets Summary (whilst Bowling)}
\usage{
getWicketsSummary(PlayerID, MatchType)
}
\arguments{
\item{PlayerID}{ESPNCricinfo Player ID.}

\item{MatchType}{Type of Match Played (1 for Test; 2 for ODI; 3 for T20I ; 11 for All)}
}
\value{
Returns DataFrame containing complete Wickets Summary of a Bowler
}
\description{
This function scraps exclusive dismissal summary (as a bowler) info from Statsguru and returns a dataframe
}
\examples{
sachin = getWicketsSummary(35320,11)
}
