% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting_dismissals.r
\name{split_DismissalSummary_Batting}
\alias{split_DismissalSummary_Batting}
\title{Splits Dismissal Summary data into respective categories}
\usage{
split_DismissalSummary_Batting(dataframe)
}
\arguments{
\item{dataframe}{Output of getDismissalSummary().}
}
\value{
Returns a list comprising data segregated into different dismissal categories
}
\description{
This function takes in output dataframe of getDismissalSummary and returns a list comprising data segregated into different dismissal categories.
}
\examples{
sachin_DismissalSummary_Batting = statsguRu::get_DismissalSummary_Batting(PlayerID = 35320, MatchType = 2)
statsguRu::split_DismissalSummary_Batting(sachin_DismissalSummary_Batting)
}
\keyword{Batting}
