% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowling_Dismissals.R
\name{split_DismissalSummary_Bowling}
\alias{split_DismissalSummary_Bowling}
\title{Splits Wickets Summary data into different categories}
\usage{
split_DismissalSummary_Bowling(data, MatchType)
}
\arguments{
\item{data}{Output of statsguRu::get_DismissalSummary_Bowling().}

\item{MatchType}{Type of Match Played (1 for Test; 2 for ODI; 3 for T20I ; 11 for All)}
}
\value{
Returns a list comprising data segregated into different categories of Wickets Summary statistics
}
\description{
This function takes in "Wickets Summary" dataframe  and returns a list comprising data segregated into different categories that can inferred from a Players's Bowling Summary.
}
\examples{
ashwin_DismissalSummary_Bowling = statsguRu::get_DismissalSummary_Bowling(PlayerID = 26421, MatchType = 1)
statsguRu::split_DismissalSummary_Bowling(data = ashwin_DismissalSummary_Bowling, MatchType = 1)
}
\keyword{Bowling}
