% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{split_Summary_Batting}
\alias{split_Summary_Batting}
\title{Splits Batting Summary data into different categories}
\usage{
split_Summary_Batting(data, MatchType)
}
\arguments{
\item{data}{Output of statsguRu::get_Summary_Batting().}

\item{MatchType}{Type of Match Played (1 for Test; 2 for ODI; 3 for T20I ; 11 for All)}
}
\value{
Returns a list comprising data segregated into different categories of Batting Summary statistics
}
\description{
This function takes in "Batting Summary" dataframe  and returns a list comprising data segregated into different categories that can inferred from a Players's Batting Summary.
}
\examples{
sachin_BattingSummary = statsguRu::get_Summary_Batting(PlayerID = 35320, MatchType = 2)
statsguRu::split_Summary_Batting(data = sachin_BattingSummary, MatchType = 2)
}
\keyword{Batting}
