% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowling.R
\name{split_Summary_Bowling}
\alias{split_Summary_Bowling}
\title{Splits Bowling Summary data into different categories}
\usage{
split_Summary_Bowling(data, MatchType)
}
\arguments{
\item{data}{Output of statsguRu::get_Summary_Bowling().}

\item{MatchType}{Type of Match Played (1 for Test; 2 for ODI; 3 for T20I ; 11 for All)}
}
\value{
Returns a list comprising data segregated into different categories of Bowling Summary statistics
}
\description{
This function takes in "Bowling Summary" dataframe  and returns a list comprising data segregated into different categories that can inferred from a Players's Bowling Summary.
}
\examples{
ashwin_BowlingSummary = statsguRu::get_Summary_Bowling(PlayerID = 26421, MatchType = 1)
statsguRu::split_Summary_Bowling(data = ashwin_BowlingSummary, MatchType = 1)
}
\keyword{Bowling}
