% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{splitBattingSummary}
\alias{splitBattingSummary}
\title{Splits Batting Summary data into different categories}
\usage{
splitBattingSummary(data)
}
\arguments{
\item{data}{Output of getBattingSummary().}
}
\value{
Returns a list comprising data segregated into different categories of Batting Summary statistics
}
\description{
This function takes in "Batting Summary" dataframe  and returns a list comprising data segregated into different categories that can inferred from a Players's Batting Summary.
}
\examples{
sachin = getBattingSummary(35320,11)
sachin1 = splitBattingSummary(sachin)
}
