% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismissal.r
\name{splitDismissalSummary}
\alias{splitDismissalSummary}
\title{Splits Dismissal Summary data into respective categories}
\usage{
splitDismissalSummary(dataframe)
}
\arguments{
\item{dataframe}{Output of getDismissalSummary().}
}
\value{
Returns a list comprising data segregated into different dismissal categories
}
\description{
This function takes in output dataframe of getDismissalSummary and returns a list comprising data segregated into different dismissal categories.
}
\examples{
sachin = getDismissalSummary(35320,11)
sachin1 = splitDismissalSummary(sachin)
}
