% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{dispBattingAveByContinent}
\alias{dispBattingAveByContinent}
\title{Displays Barplot of Players's Batting Average by List of Continents Played by the Player}
\usage{
dispBattingAveByContinent(data)
}
\arguments{
\item{data}{Output of splitBattingSummary.}
}
\value{
Plots a player's batting average in every contient.
}
\description{
This function takes in "Batting Summary" dataframe modified after splitBattingSummary and plots a player's batting average in every contitent.
}
\examples{
sachin = getBattingSummary(35320,1)
sachin1 = splitBattingSummary(sachin,1)
dispBattingAveByContinent(sachin1)
}
