% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{dispBattingAveByHostCountry}
\alias{dispBattingAveByHostCountry}
\title{Displays Barplot of Players's Batting Average by List of Host Country}
\usage{
dispBattingAveByHostCountry(
  data,
  plot_title = "Batting Average in Host Country"
)
}
\arguments{
\item{data}{Output of splitBattingSummary.}
}
\value{
Plots a player's batting average in Every host Country.
}
\description{
This function takes in "Batting Summary" dataframe modified after splitBattingSummary and plots a player's batting average in every host opposition team.
}
\examples{
sachin_BattingSummary = statsguRu::get_Summary_Batting(PlayerID = 35320, MatchType = 2)
sachin_BattingSummary = statsguRu::split_Summary_Batting(sachin_BattingSummary, MatchType = 2)
statsguRu::dispBattingAveByHostCountry(data = sachin_BattingSummary)
}
\keyword{Batting}
