% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{dispBattingAveByHostCountry}
\alias{dispBattingAveByHostCountry}
\title{Displays Barplot of Players's Batting Average by List of Host Country}
\usage{
dispBattingAveByHostCountry(data)
}
\arguments{
\item{data}{Output of splitBattingSummary.}
}
\value{
Plots a player's batting average in Every host Country.
}
\description{
This function takes in "Batting Summary" dataframe modified after splitBattingSummary and plots a player's batting average in every host opposition team.
}
\examples{
sachin = getBattingSummary(35320,11)
sachin1 = splitBattingSummary(sachin,11)
dispBattingAveByHostCountry(sachin1)
}
