% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{dispBattingAveByOpposition}
\alias{dispBattingAveByOpposition}
\title{Displays Barplot of Players's Batting Average by List of Oppostion teams}
\usage{
dispBattingAveByOpposition(data)
}
\arguments{
\item{data}{Output of splitBattingSummary.}
}
\value{
Plots a player's batting average against every opposition team.
}
\description{
This function takes in "Batting Summary" dataframe modified after splitBattingSummary and plots a player's batting average against every opposition team.
}
\examples{
sachin = getBattingSummary(35320,2)
sachin1 = splitBattingSummary(sachin,2)
dispBattingAveByOpposition(sachin1)
}
