% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{dispBattingAveByPosPlayed}
\alias{dispBattingAveByPosPlayed}
\title{Displays Barplot of Players's Batting according to the Player's Batting Position}
\usage{
dispBattingAveByPosPlayed(data)
}
\arguments{
\item{dataframe}{Output of splitBattingSummary.}
}
\value{
Plots a player's batting position-wise batting average.
}
\description{
This function takes in "Batting Summary" dataframe modified after splitBattingSummary and plots a player's batting position-wise batting average.
}
\examples{
sachin = getBattingSummary(35320,11)
sachin1 = splitBattingSummary(sachin1)
dispBattingAveByPosPlayed(sachin1)
}
