% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{dispBattingAveByYears}
\alias{dispBattingAveByYears}
\title{Displays Barplot of Players's Batting Average Year-wise}
\usage{
dispBattingAveByYears(data)
}
\arguments{
\item{dataframe}{Output of splitBattingSummary.}
}
\value{
Plots a player's year-wise batting average.
}
\description{
This function takes in "Batting Summary" dataframe modified after splitBattingSummary and plots a player's year-wise batting average.
}
\examples{
dispBattingAveByContinent(dataframe)
}
