% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{dispBattingAveSRByOpposition}
\alias{dispBattingAveSRByOpposition}
\title{Displays Scatterplot of a Players's Batting Average and Strike rate by List of Oppostion teams}
\usage{
dispBattingAveSRByOpposition(data)
}
\arguments{
\item{data}{Output of splitBattingSummary.}
}
\value{
Plots a scatterplot pitting player's batting average and strike-rate against every opposition team.
}
\description{
This function takes in "Batting Summary" dataframe modified after splitBattingSummary and plots a player's batting average and Strike Rate against every opposition team.
}
\examples{
sachin = getBattingSummary(35320,2)
sachin1 = splitBattingSummary(sachin,2)
dispBattingAveSRByOpposition(sachin1)
}
