% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlDismissal.R
\name{dispBowlerDismissals}
\alias{dispBowlerDismissals}
\title{Plots a pie chart of Dismissal Type Summary of a Bowler.}
\usage{
dispBowlerDismissals(data)
}
\arguments{
\item{data}{Output of splitWicketsSummary().}
}
\value{
Plots a pie chart of Dismissal Type Summary of a Bowler.
}
\description{
This function takes in output of splitDismissalSummary and plots the Dismissal Type Summary of a Bowler as a pie chart.
}
\examples{
sachin = getWicketsSummary(35320,1)
sachin1 = splitWicketsSummary(sachin,1)
dispBowlerDismissals(sachin1)
}
