% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowling.R
\name{dispBowlingERByYears}
\alias{dispBowlingERByYears}
\title{Displays Barplot of Players's Year-wise Bowling Economy Rate}
\usage{
dispBowlingERByYears(data)
}
\arguments{
\item{data}{Output of splitBowlingSummary.}
}
\value{
Plots a player's year-wise Bowling Economy Rate
}
\description{
This function takes in "Batting Summary" dataframe modified after splitBowlingSummary and plots a player's Year-wise Bowling Economy Rate.
}
\examples{
sachin = getBowlingSummary(35320,11)
sachin1 = splitBowlingSummary(sachin,11)
dispBowlingERByYears(sachin1)
}
