% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowling.R
\name{dispWktsTakenAtHostCountry}
\alias{dispWktsTakenAtHostCountry}
\title{Displays Barplot of Players's Wickets by List of Host Countries}
\usage{
dispWktsTakenAtHostCountry(data)
}
\arguments{
\item{data}{Output of splitBowlingSummary.}
}
\value{
Plots a player's sum of wickets taken in every host country
}
\description{
This function takes in "Bowling Summary" dataframe modified after splitBowlingSummary and plots a player's sum of wickets taken against every opposition team.
}
\examples{
sachin = getBowlingSummary(35320,2)
sachin1 = splitBowlingSummary(sachin,2)
dispWktsTakenAtHostCountry(sachin1)
}
