% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlDismissal.R
\name{splitWicketsSummary}
\alias{splitWicketsSummary}
\title{Splits Wickets Summary data into different categories}
\usage{
splitWicketsSummary(data, MatchType)
}
\arguments{
\item{data}{Output of getWicketsSummary().}

\item{MatchType}{Type of Match Played (1 for Test; 2 for ODI; 3 for T20I ; 11 for All)}
}
\value{
Returns a list comprising data segregated into different categories of Wickets Summary statistics
}
\description{
This function takes in "Wickets Summary" dataframe  and returns a list comprising data segregated into different categories that can inferred from a Players's Bowling Summary.
}
\examples{
sachin = getWicketsSummary(35320,1)
sachin1 = splitWicketsSummary(sachin,1)
dhoni = getWicketsSummary(28081,2)
dhoni1 = splitWicketsSummary(dhoni,2)
kohli = getWicketsSummary(253802,3)
kohli1 = splitWicketsSummary(kohli,3)
}
